+++ src/components/output.tsx
@@ -1,20 +1,16 @@
-import { marked } from "marked"
 import { useContext } from "react"
 
 import { OutputContext } from "~lib/context"
 
+import { MarkdownComponent } from "./makrdown"
+
 export function OutputComponent() {
-  const { isFullMode, responseMessage } = useContext(OutputContext)
+  const { responseMessage } = useContext(OutputContext)
 
   if (responseMessage) {
-    const html = marked(responseMessage)
     return (
       <div className="box-border px-4 mt-4 overflow-auto ">
-        <article
-          className={`${
-            isFullMode ? "prose-base" : "prose-xs"
-          } dark:text-white`}
-          dangerouslySetInnerHTML={{ __html: html }}></article>
+        <MarkdownComponent text={responseMessage} />
       </div>
     )
   }