+++ src/lib/api/chatgpt-web.ts
@@ -1,6 +1,9 @@
 import { v4 as uuidv4 } from "uuid"
 
+import type { PlasmoMessaging } from "@plasmohq/messaging"
+
 import { storage } from "~lib/storage"
+import { parseSSEResponse } from "~lib/utils/sse"
 
 const CHATGPT_MODEL = "text-davinci-002-render-sha"
 const CHATGPT_HOST = "https://chat.openai.com"
@@ -26,7 +29,10 @@ async function getAccessToken(): Promise<string> {
   return data.accessToken
 }
 
-async function PostChatGPT(prompt: string): Promise<string> {
+async function PostChatGPTStream(
+  prompt: string,
+  res: PlasmoMessaging.Response<any>
+) {
   const accessToken = await getAccessToken()
 
   const cacheConversationId = await storage.get(CACHE_KEY_CONVERSATION_ID)
@@ -36,21 +42,17 @@ async function PostChatGPT(prompt: string): Promise<string> {
 
   const data = {
     action: "next",
-    conversation_id: uuidv4(),
+    conversation_id: undefined,
     messages: [
       {
-        author: {
-          role: "user"
-        },
+        author: { role: "user" },
         id: uuidv4(),
-        role: "user",
         content: { content_type: "text", parts: [prompt] }
       }
     ],
     parent_message_id: cacheConversationId,
     model: CHATGPT_MODEL
   }
-  // console.log(`ChatGPTWeb request: ${JSON.stringify(data)}`)
   const url = `${CHATGPT_HOST}/backend-api/conversation`
   const resp = await fetch(url, {
     method: "POST",
@@ -61,30 +63,51 @@ async function PostChatGPT(prompt: string): Promise<string> {
     body: JSON.stringify(data)
   })
 
+  let conversationId: string = ""
+
   if (resp.status == 200) {
-    return parseChatGPTResponse(await resp.text())
-  }
-  if (resp.status == 403 || resp.status == 401) {
-    await storage.remove(CACHE_KEY_TOKEN)
-    return "ChatGPT return 403 FORBIDDEN, please login in to https://chat.openai.com"
+    await parseSSEResponse(resp, (message) => {
+      if (message === "[DONE]") {
+        // console.debug("chatgpt sse message done, start remove conversation")
+        removeConversation(conversationId)
+        return
+      }
+      try {
+        const data = JSON.parse(message)
+        const text = data.message?.content?.parts?.[0]
+        if (text) {
+          // console.debug("chatgpt sse message", text)
+          res.send(text)
+          conversationId = data.conversation_id
+        }
+      } catch (err) {
+        console.error(err)
+        res.send(err)
+        return
+      }
+    })
   } else {
-    console.log(`fail: ${resp.text}`)
-    return `ChatGPT return error, please retry, status: ${resp.status}, error: ${resp.text}`
+    res.send(resp.statusText)
   }
 }
 
-function parseChatGPTResponse(data: string) {
-  const texts = data.split("\n")
-  let res = ""
-  for (let text of texts) {
-    try {
-      const tmp = JSON.parse(text.slice(6))
-      res = tmp.message.content.parts[0]
-    } catch (error) {
-      // console.log(`parse error, data: ${text} error: ${error}`);
-    }
+async function removeConversation(id: string) {
+  const accessToken = await getAccessToken()
+  try {
+    const resp = await fetch(`${CHATGPT_HOST}/backend-api/conversation/${id}`, {
+      method: "PATCH",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${accessToken}`
+      },
+      body: JSON.stringify({
+        is_visible: false
+      })
+    })
+    console.log(await resp.json())
+  } catch (err) {
+    console.error(err)
   }
-  return res
 }
 
-export { PostChatGPT }
+export { PostChatGPTStream }